+++ auth/src/main/java/com/business/auth/infrastructure/dto/UserResponse.java
@@ -1,4 +1,4 @@
-package com.business.auth.infrastructure.client.dto;
+package com.business.auth.infrastructure.dto;
 
 import lombok.Getter;
 // User 서비스 → Auth 서비스로 유저 정보를 응답할 때 사용되는 DTO
